requirements.txt
fastapi==0.95.1
line-bot-sdk==2.4.2
pydantic==1.10.7
python-dotenv==1.0.0
python-multipart==0.0.5
uvicorn==0.15.0
